fix: navigation ok
